Split tools and commons
Commons are necessary to run core functions, tools are not
