+++ openfisca_core/formulas.py
@@ -19,7 +19,7 @@ from .base_functions import (
     requested_period_last_or_next_value,
     requested_period_last_value,
     )
-from .tools import empty_clone, stringify_array
+from .commons import empty_clone, stringify_array
 
 
 log = logging.getLogger(__name__)
