+++ openfisca_core/holders.py
@@ -6,7 +6,7 @@ from __future__ import division
 import numpy as np
 
 from . import periods
-from .tools import empty_clone
+from .commons import empty_clone
 
 
 class DatedHolder(object):
